servico-transcricao/backend/requirements.txt
fastapi
werkzeug
PyMuPDF
openai-whisper
numpy
librosa
soundfile
noisereduce
pydub
python-jose[cryptography]
slowapi 
passlib
bcrypt<4.0
deepface
facenet-pytorch
tf-keras
python-multipart
